Test/requirements.txt
aniso8601==8.1.1
appdirs==1.4.4
attrs==20.3.0
base58==2.1.0
bcrypt==3.2.0
bitarray==1.2.2
blinker==1.4
certifi==2020.11.8
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cycler==0.10.0
cytoolz==0.11.0
distlib==0.3.1
dominate==2.6.0
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
flask-blueprint==1.3.0
Flask-Bootstrap==3.3.7.1
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
hexbytes==0.2.1
idna==2.10
iniconfig==1.1.1
ipfshttpclient==0.7.0a1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
kiwisolver==1.3.1
lru-dict==1.1.7
MarkupSafe==1.1.1
matplotlib==3.3.4
mpld3==0.5.2
multiaddr==0.0.9
netaddr==0.8.0
numpy==1.20.0
packaging==20.9
pandas==1.2.1
parsimonious==0.8.1
Pillow==8.1.0
pipenv==2020.11.15
pluggy==0.13.1
protobuf==3.14.0
py==1.10.0
py-solc-x==1.1.0
pycparser==2.20
pycryptodome==3.9.9
PyJWT==1.4.2
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rlp==2.0.1
scipy==1.6.0
seaborn==0.11.1
semantic-version==2.8.5
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
toolz==0.11.1
urllib3==1.26.3
uWSGI==2.0.19.1
varint==1.0.2
virtualenv==20.2.1
virtualenv-clone==0.5.4
visitor==0.1.3
web3==5.16.0
websockets==8.1
Werkzeug==1.0.1
WTForms==2.3.3
